[fix ~ tools..3.c ] (fix typins on functions)

+++ headers/garbage_collection.h
@@ -2,10 +2,19 @@
 #define GARBAGE_COLLETION
 /* Garbage Collection*/
 /* Garbage Collector*/
+/**
+ * struct _garbage - struct
+ * @subscriptions: subscriptor
+ * @malloc: malloc
+ * @free_all: free
+ * @malloc_acumulator: acumulator
+ */
 typedef struct _garbage
 {
 	void **subscriptions;
-	int malloc_acumulator;
+
+	int long malloc_acumulator;
+
 	void *(*malloc)(void *garbage_collector_t, size_t size);
 	void *(*free_all)(void *garbage_collector_t);
 } garbage_collector_t;
